+++ packages/core/src/index.js
@@ -1,0 +1,8 @@
+//--------------------------------------------------------
+//-- Prettier config
+//--------------------------------------------------------
+import loader from "@absolunet/prettier-loader";
+
+module.exports = loader({
+	root: __dirname,
+});