+++ packages/pwastudio/src/index.js
@@ -1,0 +1,10 @@
+//--------------------------------------------------------
+//-- Prettier config
+//--------------------------------------------------------
+import coreConfig from "@absolunet/prettier-config-core";
+import loader from "@absolunet/prettier-loader";
+
+module.exports = loader({
+	root: __dirname,
+	parentConfig: coreConfig,
+});