+++ CHANGELOG.md
@@ -13,14 +13,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-## [0.1.0] - 9999-01-01
+## [1.0.0] - 2021-09-30
+### Changed
+- Official release
+
+
+
+## [0.2.1] - 2021-09-28
+### Changed
+- Update to Prettier 2.4.1
+- Maintenance updates
+
+
+
+## [0.2.0] - 2021-09-14
+### Changed
+- Update to Prettier 2.4.0
+- Rename `jsxBracketSameLine` to `bracketSameLine`
+- Manager and tester update
+
+
+
+## [0.1.0] - 2021-08-18
 ### Added
-- Initial
+- Add `loader` package
+- Add `core` config
+- Add `pwastudio` config
+- Add `spire` config
 
 
 
 
 
 
-[Unreleased]: https://github.com/absolunet/prettier-config/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/absolunet/prettier-config/compare/1.0.0...HEAD
+[1.0.0]:      https://github.com/absolunet/prettier-config/compare/0.2.1...1.0.0
+[0.2.1]:      https://github.com/absolunet/prettier-config/compare/0.2.0...0.2.1
+[0.2.0]:      https://github.com/absolunet/prettier-config/compare/0.1.0...0.2.0
 [0.1.0]:      https://github.com/absolunet/prettier-config/releases/tag/0.1.0